// packages/sensinact-mapping-wizard/src/generated/Measurement.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EDataType, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';
import type { FeaturePath } from './FeaturePath';
import type { StoragePreset } from './StoragePreset';
import type { RetentionPreset } from './RetentionPreset';

/**
 * Measurement
 * @generated
 */
export interface Measurement extends EObject {
  selected?: boolean;
  valuePath: FeaturePath;
  label?: string;
  unit?: string;
  unitPath?: FeaturePath;
  serviceGroup?: string;
  targetType?: EDataType;
  storagePreset?: StoragePreset;
  retentionPreset?: RetentionPreset;

}