// packages/sensinact-mapping-wizard/src/generated/FriendlyNameSource.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

/**
 * FriendlyNameSource
 * @generated
 */
export const FriendlyNameSource = {
  NONE: 'NONE',
  FROM_FIELD: 'FROM_FIELD',
  STATIC: 'STATIC'
} as const;

export type FriendlyNameSource = typeof FriendlyNameSource[keyof typeof FriendlyNameSource];

/**
 * Get enum value by name
 */
export function getFriendlyNameSource(name: string): FriendlyNameSource | undefined {
  return (FriendlyNameSource as Record<string, string>)[name] as FriendlyNameSource | undefined;
}

/**
 * Get enum value by literal
 */
export function getFriendlyNameSourceByLiteral(literal: string): FriendlyNameSource | undefined {
  if (literal === 'NONE') return FriendlyNameSource.NONE;
  if (literal === 'FROM_FIELD') return FriendlyNameSource.FROM_FIELD;
  if (literal === 'STATIC') return FriendlyNameSource.STATIC;
  return undefined;
}